refactor: move "createCalendar" into utils
