+++ packages/osc-ui/src/components/DatePicker/Calendar/DecadeCalendar.tsx
@@ -11,8 +11,7 @@ import React, { useEffect, useState } from 'react';
 import { classNames } from '../../../utils/classNames';
 import { Button } from '../../Button/Button';
 import { Icon } from '../../Icon/Icon';
-import { checkYearRange, selectDateHandler, setDisabledRange } from '../utils';
-import { createCalendar } from './Calendar';
+import { checkYearRange, createCalendar, selectDateHandler, setDisabledRange } from '../utils';
 
 type Year = {
     isDisabled: boolean;
