+++ packages/osc-ui/src/components/DatePicker/DateField/DateField.tsx
@@ -1,8 +1,8 @@
-import { GregorianCalendar } from '@internationalized/date';
 import type { AriaButtonProps } from '@react-aria/button';
 import type { AriaDateFieldProps } from '@react-aria/datepicker';
 import { useDateField, useDateSegment } from '@react-aria/datepicker';
 import { useLocale } from '@react-aria/i18n';
+import type { DateFieldState, DateSegment } from '@react-stately/datepicker';
 import { useDateFieldState } from '@react-stately/datepicker';
 import type { DateValue } from '@react-types/calendar';
 import React, { useRef } from 'react';
@@ -10,20 +10,7 @@ import { useUniqueId } from '../../../hooks/useUniqueId';
 import { Icon } from '../../Icon/Icon';
 import '../date-field.scss';
 import { ReactAriaButton } from '../ReactAriaComponents/ReactAriaComponents';
-import type { DateFieldState, DateSegment } from '@react-stately/datepicker';
-
-// Ensure that we just use the Gregorian Calendar to limit package size.
-// If needed more can be added in at a later point. See below link for example:
-// https://react-spectrum.adobe.com/react-aria/useDateField.html#reducing-bundle-size
-const createCalendar = (identifier) => {
-    switch (identifier) {
-        case 'gregory':
-            return new GregorianCalendar();
-        default:
-            throw new Error(`Unsupported calendar ${identifier}`);
-    }
-};
-
+import { createCalendar } from '../utils';
 interface DateFieldProps extends AriaDateFieldProps<DateValue> {
     buttonProps: AriaButtonProps;
 }
